Improved coding framework and added options to change brain area, making trees with different numbers of genes
Made DT_classification_multiple.m into a function where we can choose num genes to use in each tree, and what area is the target. Can also choose a number of samples < cols to execute the algorithm on for testing with less runtime.
The function also generates a histogram of accuracies every time it is run.
Made new script DT_general to run the other functions at a high level.
Modified the rest of the functions somewhat to fit the requirements, e.g. made area an input of filter_nans.m but eventually this should be managed better in an outer function.